# App/BackEnd/Conduit.API/src/Conduit.API/wwwroot/openapi-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "contact" : {
      "name" : "RealWorld",
      "url" : "https://realworld.io"
    },
    "description" : "Conduit API",
    "license" : {
      "name" : "MIT License",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "title" : "Conduit API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "/api"
  } ],
  "paths" : {
...
    "/profiles/{username}" : {
      "get" : {
        "description" : "Get a profile of a user of the system. Auth is optional",
        "operationId" : "GetProfileByUsername",
        "parameters" : [ {
          "description" : "Username of the profile to get",
          "explode" : false,
          "in" : "path",
          "name" : "username",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : { },
            "description" : "Unauthorized"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            },
            "description" : "Unexpected error"
          }
        },
        "summary" : "Get a profile",
        "tags" : [ "Profile" ]
      }
    },
    "/profiles/{username}/follow" : {
      "delete" : {
        "description" : "Unfollow a user by username",
        "operationId" : "UnfollowUserByUsername",
        "parameters" : [ {
          "description" : "Username of the profile you want to unfollow",
          "explode" : false,
          "in" : "path",
          "name" : "username",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : { },
            "description" : "Unauthorized"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            },
            "description" : "Unexpected error"
          }
        },
        "security" : [ {
          "Token" : [ ]
        } ],
        "summary" : "Unfollow a user",
        "tags" : [ "Profile" ]
      },
      "post" : {
        "description" : "Follow a user by username",
        "operationId" : "FollowUserByUsername",
        "parameters" : [ {
          "description" : "Username of the profile you want to follow",
          "explode" : false,
          "in" : "path",
          "name" : "username",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : { },
            "description" : "Unauthorized"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            },
            "description" : "Unexpected error"
          }
        },
        "security" : [ {
          "Token" : [ ]
        } ],
        "summary" : "Follow a user",
        "tags" : [ "Profile" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ProfileResponse" : {
        "example" : {
          "profile" : {
            "image" : "image",
            "following" : true,
            "bio" : "bio",
            "username" : "username"
          }
        },
        "properties" : {
          "profile" : {
            "$ref" : "#/components/schemas/Profile"
          }
        },
        "required" : [ "profile" ],
        "type" : "object"
      },
      "Profile" : {
        "example" : {
          "image" : "image",
          "following" : true,
          "bio" : "bio",
          "username" : "username"
        },
        "properties" : {
          "username" : {
            "type" : "string"
          },
          "bio" : {
            "type" : "string"
          },
          "image" : {
            "type" : "string"
          },
          "following" : {
            "type" : "boolean"
          }
        },
        "required" : [ "bio", "following", "image", "username" ],
        "type" : "object"
      },
...
      "GenericErrorModel" : {
        "properties" : {
          "errors" : {
            "$ref" : "#/components/schemas/GenericErrorModel_errors"
          }
        },
        "required" : [ "errors" ],
        "type" : "object"
      },
      "GenericErrorModel_errors" : {
        "properties" : {
          "body" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "required" : [ "body" ],
        "type" : "object"
      }
    },
    "securitySchemes" : {
      "Token" : {
        "description" : "For accessing the protected API resources, you must have received a a valid JWT token after registering or logging in. This JWT token must then be used for all protected resources by passing it in via the 'Authorization' header.\n\nA JWT token is generated by the API by either registering via /users or logging in via /users/login.\n\nThe following format must be in the 'Authorization' header :\n\n    Token xxxxxx.yyyyyyy.zzzzzz\n    \n",
        "in" : "header",
        "name" : "Authorization",
        "type" : "apiKey"
      }
    }
  }
}